refactor: Add PhotoService
